Check for xdebug before disabling in ci/prepare.sh.

+++ ci/prepare.sh
@@ -7,7 +7,11 @@ set -ex
 WP_CLI_BIN_DIR=${WP_CLI_BIN_DIR-/tmp/wp-cli-phar}
 
 # Disable XDebug to speed up Composer and test suites.
-phpenv config-rm xdebug.ini
+if [ -f ~/.phpenv/versions/$(phpenv version-name)/etc/conf.d/xdebug.ini ]; then
+  phpenv config-rm xdebug.ini
+else
+  echo "xdebug.ini does not exist"
+fi
 
 composer install --no-interaction --prefer-source
